# BE/myapp/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/events/create": {
      "get": {
        "tags": ["Admin API"],
        "summary": "근무 일정 생성",
        "description": "CreateSchedule.js - adminEvents.js",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Admin API"],
        "summary": "근무 일정 생성",
        "description": "CreateSchedule.js - adminEvents.js",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "start": {
                  "example": "any"
                },
                "end": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "allDay": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/edit/{id}": {
      "get": {
        "tags": ["Admin API"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Admin API"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "start": {
                  "example": "any"
                },
                "end": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "allDay": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/delete/{id}": {
      "post": {
        "tags": ["Admin API"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/all": {
      "get": {
        "tags": ["Admin API"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/requests": {
      "get": {
        "tags": ["Admin API"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/approve": {
      "post": {
        "tags": ["Admin API"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/approve/{requestId}": {
      "post": {
        "tags": ["Admin API"],
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/events/reject/{requestId}": {
      "post": {
        "tags": ["Admin API"],
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/worker/events/apply": {
      "post": {
        "tags": ["Worker API"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workerId": {
                  "example": "any"
                },
                "userName": {
                  "example": "any"
                },
                "start": {
                  "example": "any"
                },
                "end": {
                  "example": "any"
                },
                "lastShiftStart": {
                  "example": "any"
                },
                "lastShiftEnd": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}